requirements.txt
python-telegram-bot
aiohttp
gspread
google-auth
